feat: Implement delete post functionality and improve error handling in post creation

+++ src/handlers/postsHandler.ts
@@ -1,10 +1,10 @@
-import { Request, Response } from "express";
+import e, { Request, Response } from "express";
 import Post from "../models/Posts.model";
 import User from "../models/Users.model";
 import Like from "../models/Likes.model";
 import Comment from "../models/Comments.model";
 import { cloudinaryOptions } from "../config/cloudinary";
-import fs from 'fs'
+import fs from "fs";
 
 export const getAllPosts = async (req: Request, res: Response) => {
   const posts = await Post.findAll();
@@ -55,27 +55,25 @@ export const createNewPost = async (req: Request, res: Response) => {
     const { id_user } = req.userData;
 
     if (!title || !content) {
-      return res.status(400).json('All fields must be filled')
+      return res.status(400).json("All fields must be filled");
     }
-     
+
     if (req.file) {
-       const uploadResult = await cloudinaryOptions.uploader.upload(
+      const uploadResult = await cloudinaryOptions.uploader.upload(
         req.file.path,
         {
           folder: "posts",
         }
-        
       );
 
-      imageURL  = uploadResult.secure_url
-
+      imageURL = uploadResult.secure_url;
     }
 
     const newPost = await Post.create({
       title: title,
       content: content,
       image: imageURL,
-      user_id: id_user
+      user_id: id_user,
     });
 
     res.json(newPost);
@@ -86,6 +84,28 @@ export const createNewPost = async (req: Request, res: Response) => {
       error: error.message,
     });
   } finally {
-    fs.unlinkSync(req.file.path)
+    if (req.file) fs.unlinkSync(req.file.path);
   }
 };
+
+export const deletePost = async (req: Request, res: Response) => {
+  try {
+    const { id_post } = req.params;
+    const { id_user } = req.userData;
+
+    const post = await Post.findByPk(id_post);
+
+    if (post.user_id !== id_user) {
+      return res
+        .status(401)
+        .json({ message: "You are not authorized to do this action" });
+    }
+
+    await post.destroy();
+
+    res.status(200).json({ message: "Post deleted correctly" });
+  } catch (error) {
+    console.error(error);
+  }
+};
+
